Modified the SurfaceTracker so that a possible race condition is prevented. Instead of blending the new surface into the tracked surface, the tracked surface is replaced by the new surface, and then blended into it. This way the surfaceModel will not be subject to any change once it leaves the surface detection thread - i.e. a race condition is impossible.
Further, there is some uncommented code within the surfaceDetector that
allows a full deep copy of the passed surfaceModels. This code snippet
should always then be uncommented if a race condition is suspected.

These changes do not influence the flickering of the visualizer which
shows that flickering cannot be caused by race conditions.