Fix segfault when creating multiple RoboService instances Related to issue #32
This moves the creation of RobotService into the parsing step for RobotAggregator.
This allows us to pass the new RobotService directly to the RobotAggregator without
needing to worry about storing it somewhere else.

Since the RobotService is no longer held by the FileConfigParser once parsing is
complete, RobotAggregator was also updated to accept a boost::shared_ptr<RobotService>,
rather than a RobotService&, which was necessary to prevent it from being deleted
after parsing was complete.