Last-minute filtering of surfaces in RobotAggregator This allows RobotAggregator to perform a last-minute test on all outgoing surfaces in order to avoid sending parts of the ground plane to the planning system. Two parameters to configure this are now required when the Aggregator has the "surfaces" data type enabled:
  - min_surface_height: minimum vertical distance a surface should be
    from the ground in order to be passed on to the planning system
  - surface_normal_tolerance: for surfaces below the min_surface_height,
    if their normal deviates from the world vertical axis (0, 0, 1) by
more than this value (in radians), then we do not consider the surface
to be part of the ground plane and will pass it on like normal (think a
ramp which intersects the ground plane, for example).

If a surface is determined to be part of the ground plane, then the
convex hull is replaced before the surface data is sent to QNX. The
dummy hull has points VERY far from the origin, so the planner should
never consider the resulting surface.