Fix dummy surface parameters for surfaces filtered by RobotAggregator Resolves issue #58
QNX has a hard limit on +/- 5 meters from lola's origin for all surfaces (and crashes if any points are outside this area) and computes the surface area of all incoming surfaces (which causes a crash if the surface has 0 area). This ensures dummy surfaces have some small area, and the points are all clusted near (-4.9, -4.9, -4.9).