Some improvements for GMM Segmenter behavior This includes a couple small changes which appear to have improved the stability of the GMM segmenter's results, although they still look like they could be better.
Closing in on issue #63.

The most important change was to wait until the surface detector had
processed the scene and made a good estimate of the ground plane. The
original GMM code waited for two iterations of the surface detector,
which did not help in my test, but bumping this up to 20 has improved
the overall results a lot (20 is probably excessive, though).

The reason this is so important is the first frame GMM processes will be
used to generate the first gaussian and initial distribution over the
pointcloud. If this frame is radically different from a 'typical' frame
(e.g. because it includes the ground and all future frames do not) then
the math all gets thrown off.

The other change was to prune a few additional points from each obstacle
cloud based on the logic which was used in the old GMM obstacle
approximation code. In general, the sizes of the SSVs generated by the
GMM SSV approximation do not match those generated by the
MomentofInertia approximator (usually the MomentOfInertia approximator
produces larger SSVs). This is expected since the GMM approximation is
not based on the points, themselves, but I've seen a lot of cases where
the GMM approximation was 50% smaller than the MomentOfInertia
approximation!

With this change, they are closer (I'm guessing some noisy points which
happen to be assigned to a particular Gaussian sometimes stretch the SSVs
generated by the MomentOfInertia approximator).