Allow GMM segmenter to filter outgoing obstacles by age This adds a config parameter to tell the GMM segmenter not to pass obstacles on through FrameData until they've been observed in several frames (default value is 1, which essentially means no filtering).
This should improve things a bit since real obstacles seem to have very
stable behavior, but we often see noisy obstacles popping in for just a
frame or two. The new behavior is similar to the LowPass filter's
materializing strategy, but it will NOT persist a removed obstacle for X
frames after it disappears (I don't think this should be necessary for
GMM, though).

With the way GMM currently generates obstacle IDs, it's best to keep
this behavior inside the segmenter rather than creating an extra
observer to handle it (it's possible we may delete a GMM state on the
same frame a new one is added, and the new state could end up reusing
the ID of the removed one--In the segmenter we know they're different,
but once they're passed on through FrameData they would look like the
same obstacle).